Always create snapshots
Prior to this commit, zap would only create new snapshots for filesystem
that had changed since the last snapshot.  If there were no changes,
then the name of the most recent snapshot would be updated with the
current date.  There were problems with this logic, so just
unconditionally create the snapshot.