Add send() for remote, differential replication.
Warning
-------
The format of zap snapshot names has changed.  They now include a
short host name to identify the host on which the snapshot was created.

The new format is @ZAP_<short host name>_<snapshot date>--<ttl>.

This means snapshots with the old format (without the host name) will have to be
destroyed manually.