Do not use -r with zfs send. Replicate child datasets individually.
As reported by David Samms <dsamms@nw-ds.com>, when replicating with -r, if
replication of one of the child datasets failed, subsequent checks for that
child dataset were ignored.

--- Example

% zap rep -v -r zroot/ROOT

Replication of zroot/ROOT completed, but replication of zroot/ROOT/default
failed.

% zap rep -v -r zroot/ROOT

The local and remote snapshots matched for zroot/ROOT and the snapshots for
zroot/ROOT/default were not compared, so no errors were reported.

---