Fix issue when destroying snapshot of top-level dataset
To extract the pool name from a snapshot name, a parameter expansion was
removing everything after '/'.  However, the top-level dataset contains no '/'
in its name (its name is the same as the pool's name).  This means the entire
snapshot name was being incorrectly returned for the name of the pool.

This commit fixes the parameter expansion to handle the case when the dataset
name contains no '/'.